fix: release resource even has fatal error
Originally, an immediate exit is made after a fatal error. Resources
cannot be freed gracefully under such condition. Now, we use the macro
of Bison to end the parsing routine, so resources can be released
properly. This also only ends the `yyparse`, so the file can be closed
and the `yylex_destroyed` can be called.